Add the property of jack event delay time in the DTS settings
The delay time is set for next IRQ event, not in current IRQ event.
Ex. During plug in IRQ event, the delay time is set for push button delay or
plug out delay. Because after plug in IRQ event, the next IRQ event should be
one of push button or plug out. So it names the delay-plug-out-pb.

Signed-off-by: Oder Chiou <oder_chiou@realtek.com>